backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
paramiko==3.4.0
cryptography==41.0.7
pydantic==2.5.2
pydantic-settings==2.1.0
websockets==12.0
aiofiles==23.2.1
apscheduler==3.10.4
